Implementing back propagation net

+++ src/neuralnets/Perceptron.java
@@ -4,10 +4,10 @@ import java.util.Random;
 
 public class Perceptron {
 
-	float weights[];
+	double weights[];
 	
 	public Perceptron(int numInputs) {
-		weights = new float[numInputs + 1];
+		weights = new double[numInputs + 1];
 		
 		Random rand = new Random();
 		for(int i = 0; i < weights.length; i++) {
@@ -15,7 +15,7 @@ public class Perceptron {
 		}
 	}
 	
-	float g(float x[]) {
+	double g(double x[]) {
 		float sum = 0;
 		for(int i = 0; i < weights.length; i++) {
 			sum += weights[i] * x[i];
@@ -23,14 +23,14 @@ public class Perceptron {
 		return sum;
 	}
 
-	int o(float input[]) {
+	int o(double input[]) {
 		if (g(input) > 0)
 			return 1;
 		else
 			return 0;
 	}
 
-	public void adjustWeights(float[] sum) {
+	public void adjustWeights(double[] sum) {
 		for(int i = 0; i < weights.length; i++) {
 			weights[i] += sum[i];
 		}
